requirements.txt
musicxmatch_api==1.0.7
spotipy==2.25.1
lrclibapi==0.3.1
cryptography==45.0.6
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0